renamed aux_utils to disk_utils
