+++ tests/range_search_disk_index.cpp
@@ -8,8 +8,8 @@
 #include <set>
 #include <boost/program_options.hpp>
 
-#include "aux_utils.h"
 #include "index.h"
+#include "disk_utils.h"
 #include "math_utils.h"
 #include "memory_mapper.h"
 #include "pq_flash_index.h"
