+++ tests/search_disk_index.cpp
@@ -11,13 +11,13 @@
 #include <time.h>
 #include <boost/program_options.hpp>
 
-#include "aux_utils.h"
 #include "index.h"
+#include "utils.h"
+#include "disk_utils.h"
 #include "math_utils.h"
 #include "memory_mapper.h"
 #include "partition.h"
 #include "timer.h"
-#include "utils.h"
 #include "percentile_stats.h"
 
 #ifndef _WINDOWS
