+++ tests/search_memory_index_dynamic.cpp
@@ -17,7 +17,6 @@
 #include <unistd.h>
 #endif
 
-#include "aux_utils.h"
 #include "index.h"
 #include "memory_mapper.h"
 #include "utils.h"
