+++ tests/utils/calculate_recall.cpp
@@ -9,8 +9,8 @@
 #include <string>
 #include <vector>
 
-#include "aux_utils.h"
 #include "utils.h"
+#include "disk_utils.h"
 
 int main(int argc, char** argv) {
   if (argc != 4) {
